requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
alembic
psycopg2-binary
databases
pydantic
redis
aioredis
python-dotenv
celery[redis]
requests
